backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
redis
pyjwt
pydantic
pydantic-settings
